[Gold V] Title: 다각형의 면적, Time: 8 ms, Memory: 2152 KB -BaekjoonHub

+++ 백준/Gold/2166. 다각형의 면적/다각형의 면적.cc
@@ -3,13 +3,13 @@ using namespace std;
 
 vector<pair<int, int>> tri;
 
-long double cal(int a) {
-    long double x1 = tri[0].first;
-    long double y1 = tri[0].second;
-    long double x2 = tri[a].first;
-    long double y2 = tri[a].second;
-    long double x3 = tri[a + 1].first;
-    long double y3 = tri[a + 1].second;
+double cal(int a) {
+    double x1 = tri[0].first;
+    double y1 = tri[0].second;
+    double x2 = tri[a].first;
+    double y2 = tri[a].second;
+    double x3 = tri[a + 1].first;
+    double y3 = tri[a + 1].second;
 
     return ((x1 * y2 + x2 * y3 + x3 * y1) - (x1 * y3 + x2 * y1 + x3 * y2)) / 2;
 }
@@ -24,7 +24,7 @@ int main() {
         tri.push_back({x, y});
     }
 
-    long double res = 0;
+    double res = 0;
     for (int i = 1; i < n - 1; i++)
         res += cal(i);
